Fixed routing in index.js and removed App.js

+++ api/schema.sql
@@ -13,7 +13,7 @@ CREATE TABLE products (
     name VARCHAR(255) NOT NULL,
     description VARCHAR(255),
     price DECIMAL(10,2) NOT NULL,
-    quantity DECIMAL(10,2) NOT NULL,
+    quantity INT NOT NULL,
     PRIMARY KEY (id)
 );
